requirements.txt
python-can[serial]
ttkthemes